+++ lib/providers/userReview.dart
@@ -6,7 +6,7 @@ class ReviewDetails with ChangeNotifier {
   String userId;
   String username;
   String review;
-  int rating;
+  double rating;
   DateTime date;
   ReviewDetails(
       {required this.id,
@@ -20,18 +20,3 @@ class ReviewDetails with ChangeNotifier {
     return review;
   }
 }
-
-// class UserReview with ChangeNotifier {
-//   final String userId;
-//   final String authToken;
-//   List<reviewDetails> _reviews = [];
-//   UserReview(this.userId, this.authToken, this._reviews);
-
-//   List<reviewDetails> get reviews {
-//     return _reviews;
-//   }
-
-//   Future<void> addReview(String review, int rating, String hotelId) async {
-   
-//   }
-// }
